Fix reset of smoother

+++ local_planner/src/nodes/local_planner.cpp
@@ -59,8 +59,10 @@ void LocalPlanner::dynamicReconfigureSetParams(
   min_cloud_size_ = config.min_cloud_size_;
   min_realsense_dist_ = static_cast<float>(config.min_realsense_dist_);
   min_dist_backoff_ = static_cast<float>(config.min_dist_backoff_);
-  pointcloud_timeout_hover_ = config.pointcloud_timeout_hover_;
-  pointcloud_timeout_land_ = config.pointcloud_timeout_land_;
+  pointcloud_timeout_hover_ =
+      static_cast<float>(config.pointcloud_timeout_hover_);
+  pointcloud_timeout_land_ =
+      static_cast<float>(config.pointcloud_timeout_land_);
   children_per_node_ = config.children_per_node_;
   n_expanded_nodes_ = config.n_expanded_nodes_;
